% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_lognormal.R
\name{discrete_lognormal}
\alias{discrete_lognormal}
\title{Discrete lognormal distribution}
\usage{
discrete_lognormal(meanlog, sdlog, breaks, dim = NULL)
}
\arguments{
\item{meanlog}{unconstrained parameters giving the mean of the distribution
on the log scale}

\item{sdlog}{unconstrained parameters giving the standard deviation of the
distribution on the log scale}

\item{breaks}{a vector of breaks; observations will be treated as censored
within those breaks}

\item{dim}{a scalar giving the number of rows in the resulting greta array}
}
\description{
a discretised lognormal distribution (i.e. sampled by applying
the round operation to samples from a lognormal). Due to the numerical
instability of integrating across the distribution, a vector of breaks
must be defined and the observations will be treated as censored
within those breaks
}
