% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_normal.R
\name{discrete_normal}
\alias{discrete_normal}
\title{Discrete normal distribution}
\usage{
discrete_normal(mean, sd, breaks, edges, dim = NULL)
}
\arguments{
\item{mean}{unconstrained parameters giving the mean of the distribution}

\item{sd}{unconstrained parameters giving the standard deviation of the
distribution}

\item{breaks}{a vector of breaks; observations will be treated as censored
within those breaks}

\item{edges}{a vector of edges; length needs to be length(breaks)+1;
observations between two consecutive edges will be discretised to the
break value between the corresponding edges}

\item{dim}{a scalar giving the number of rows in the resulting greta array}
}
\description{
a discretised normal distribution (i.e. sampled by applying
the round operation to samples from a normal). Due to the numerical
instability of integrating across the distribution, a vector of breaks
must be defined and the observations will be treated as censored
within those breaks
}
