% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{greta.distributions}
\alias{greta.distributions}
\title{Extends Distributions Available in the \code{greta} package}
\description{
describe your package here, you can re-use the text from DESCRIPTION
}
\examples{

# add a simple example here to introduce the package!

}
