% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_distribution_template_test.R
\name{greta_dist_template_test}
\alias{greta_dist_template_test}
\title{Create a test template for a distribution}
\usage{
greta_dist_template_test(dist_name)
}
\arguments{
\item{dist_name}{character. A distribution name. E.g., "lognormal"}
}
\value{
text containing test template code
}
\description{
When you add a new distribution, you want to add a test to make sure it is
behaving as expected. This function generates a test template. It only
creates the text, it does not write to file. See
\code{greta_dist_write()} to write the test to file automatically.
}
\examples{
gumbel <- greta_dist_template_test("gumbel")
}
