% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-template.R
\name{greta_distribution_template}
\alias{greta_distribution_template}
\title{Create a template greta distribution}
\usage{
greta_distribution_template(dist_name, dist_arg_list)
}
\arguments{
\item{dist_name}{character length 1, name of the distribution}

\item{dist_arg_list}{character vector, arguments to the distribution}
}
\value{
text of a basic greta distribution
}
\description{
This code creates the text of a basic greta distribution. To write the text
to file, see \code{write_new_distribution()}.
}
\examples{
greta_distribution_template(
  dist_name = "lognormal",
  dist_arg_list = c("meanlog", "sdlog")
  )
}
