% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_distribution_template_test.R
\name{write_distribution_test}
\alias{write_distribution_test}
\title{Create a test template for a distribution}
\usage{
write_distribution_test(dist_name, overwrite = FALSE)
}
\arguments{
\item{dist_name}{character, name of distribution.}

\item{overwrite}{logical. Default is FALSE. Whether to overwrite the test
file if it already exists.}
}
\value{
test file written out to \code{tests/testthat/test-{dist_name}}.
}
\description{
When you add a new distribution, you want to add a test to make sure it is
behaving as expected. This function generates a test template. See also
\code{greta_distribution_template_test()} to see the text generated if you want
to save it somewhere else.
}
\examples{
\dontrun{
write_distribution_test("gumbel")
}
}
