% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-distribution.R
\name{write_greta_dist}
\alias{write_greta_dist}
\title{Write out a template greta distribution}
\usage{
write_greta_dist(dist_name, dist_arg_list)
}
\arguments{
\item{dist_name}{character length 1, name of the distribution}

\item{dist_arg_list}{character vector, arguments to the distribution}
}
\value{
text of a basic greta distribution
}
\description{
Write out a template greta distribution
}
\examples{
write_greta_dist(
  dist_name = "lognormal", 
  dist_arg_list = c("meanlog", "sdlog")
  )
}
