% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distribution.R
\name{write_new_distribution}
\alias{write_new_distribution}
\title{Write a new greta distribution template file}
\usage{
write_new_distribution(dist_name, dist_arg_list)
}
\arguments{
\item{dist_name}{name of distribution.}

\item{dist_arg_list}{arguments for distribution.}
}
\value{
writes to file a new distribution name.
}
\description{
Write a new greta distribution template file
}
\examples{
\dontrun{
write_new_distribution(
  dist_name = "lognormal", 
  dist_arg_list = c("meanlog", "sdlog")
  )
}
}
