% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated_poisson.R
\name{zero_inflated_poisson}
\alias{zero_inflated_poisson}
\title{Zero Inflated Poisson distribution}
\usage{
zero_inflated_poisson(theta, lambda, dim = NULL)
}
\arguments{
\item{theta}{proportion of zeros}

\item{lambda}{rate parameter}

\item{dim}{a scalar giving the number of rows in the resulting greta array}
}
\description{
A zero inflated poisson distribution.
}
