% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta-distributions-package.R
\docType{package}
\name{greta.distributions}
\alias{greta.distributions-package}
\alias{greta.distributions}
\title{Extends Distributions Available in the \code{greta} package}
\description{
describe your package here, you can re-use the text from DESCRIPTION
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/greta-dev/greta.distributions}
  \item Report bugs at \url{https://github.com/greta-dev/greta.distributions/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

Authors:
\itemize{
  \item Nick Golding \email{nick.golding.research@gmail.com} (\href{https://orcid.org/0000-0001-8916-5570}{ORCID})
}

}
\keyword{internal}
