% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated_poisson.R
\name{zero_inflated_poisson}
\alias{zero_inflated_poisson}
\title{Zero Inflated Poisson distribution}
\usage{
zero_inflated_poisson(lambda, pi, dim = NULL)
}
\arguments{
\item{lambda}{rate parameter}

\item{pi}{proportion of zeros}

\item{dim}{a scalar giving the number of rows in the resulting greta array}
}
\description{
A zero inflated poisson distribution.
}
\examples{
\dontrun{
zip <- zero_inflated_poisson(lambda = 2, pi = 0.2)
calculate(zip, nsim = 10)
mcmc(m)
}
}
